First improvements to _binary_op - remove creation of factory.dndarray for scalar values --> can be handled by torch.operation directly - type casting adjustments - subsequent fixes in indexing.py and tests

+++ heat/core/indexing.py
@@ -118,7 +118,7 @@ def where(cond, x=None, y=None):
             raise NotImplementedError("binary op not implemented for different split axes")
     if isinstance(x, (dndarray.DNDarray, int, float)) and isinstance(y, (dndarray.DNDarray, int, float)):
         cond = types.float(cond)
-        return (cond == 0) * y + cond * x
+        return types.float(cond == 0) * y + cond * x
     elif x is None and y is None:
         return nonzero(cond)
     else:
